+++ src/test/DruidTestService.spec.ts
@@ -3,9 +3,9 @@ import chai from "chai";
 import chaiHttp from "chai-http";
 import nock from "nock";
 import httpStatus from "http-status";
-import TestDruidQuery from "./testquery";
-import { config } from "./config";
-import constants from "../resources/constants.json";
+import { TestDruidQuery } from "./Fixtures";
+import { config } from "./Config";
+import constants from "../resources/Constants.json";
 import routes from "../routes/RoutesConfig";
 chai.should();
 chai.use(chaiHttp);
